refactor: split pattern workspace helpers
