+++ .sampo/changesets/1002-split-pattern-runtime.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split pattern workspace add helpers into focused runtime modules without changing generated output.
